About 85% Successful Breakthrough
After A LOT of research and testing over the last few days, I have some good news.
The Algorithm outputs  sorted data to a vector--from there it can be exported to a CSV, JSON, MySQL, where ever it needs to be.

The Algorithm is *mostly* reliable, however the output slightly changes based on the order it was added, I *think* that's just the nature of how it works, but I want to double check with Kelli to make sure.

That being said, the algorithm reliably sorts by the *distance* instance variables each order has and my math says each iteration is technically correct.

I have this uneasy feeling whenever I get even consistently inconsistent results in Maths, even if the inconsistencies can be predicted, it just makes me weary.

Updated with some commenting and changed the debug out statements.